pyproject.toml
[project]
name = "sfacspace-todo-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.41.0",
]
